adding activation layer
