+++ allennlp/modules/transformer/activation_layer.py
@@ -22,5 +22,5 @@ class ActivationLayer(torch.nn.Module, FromParams):
 
     def forward(self, hidden_states):
         hidden_states = self.dense(hidden_states)
-        hidden_states = self.intermediate_act_fn(hidden_states)
+        hidden_states = self.act_fn(hidden_states)
         return hidden_states
